Move v1.1 up to schema
