+++ schema/ast/ast_schema.go
@@ -15,8 +15,8 @@
 package ast // import "go.opentelemetry.io/otel/schema/v1.1/ast"
 
 import (
+	"go.opentelemetry.io/otel/schema/types"
 	ast10 "go.opentelemetry.io/otel/schema/v1.0/ast"
-	"go.opentelemetry.io/otel/schema/v1.1/types"
 )
 
 // Schema represents a Schema file in FileFormat 1.1.0 as defined in
